endpoint detail implementation

+++ src/main/java/com/mservicetech/campsite/service/CampsiteServiceImpl.java
@@ -16,6 +16,7 @@ import java.time.Duration;
 import java.time.LocalDate;
 import java.util.ArrayList;
 import java.util.List;
+import java.util.stream.Collectors;
 
 @Service
 public class CampsiteServiceImpl implements CampsiteService{
@@ -27,11 +28,8 @@ public class CampsiteServiceImpl implements CampsiteService{
 
     @Override
     public AvailableDates getAvailableDates(LocalDate startDate, LocalDate endDate) {
-        //TODO validation for input parameters
         AvailableDates availableDates = new AvailableDates();
         List<LocalDate> result = new ArrayList<>();
-        if (startDate==null) startDate = LocalDate.now();
-        if (endDate==null) endDate = LocalDate.now().plusDays(30);
         long days = Duration.between(startDate.atStartOfDay(), endDate.atStartOfDay() ).toDays();
         for (int i=1; i<=days; i++) {
             result.add(startDate.plusDays(i));
@@ -47,33 +45,38 @@ public class CampsiteServiceImpl implements CampsiteService{
     @Override
     @Transactional
     public Reservation createReservation(Reservation reservation) {
-        //TODO validation
         List<LocalDate> dateList = new ArrayList<>();
         long days = Duration.between(reservation.getArrival().atStartOfDay(), reservation.getDeparture().atStartOfDay() ).toDays();
         for (int i=0; i<=days; i++) {
             dateList.add(reservation.getArrival().plusDays(i));
         }
-        //TODO check the dates again
-        List<Client> clients = campsiteRepository.checkClientExisting(reservation.getClient());
-        if (clients.size()==0 ) {
-            long clientId = campsiteRepository.insertClient(reservation.getClient());
-            reservation.getClient().setId(clientId);
-            logger.debug("new client inserted, client id:" + clientId);
+        List<LocalDate> bookedList =  campsiteRepository.verifyDates(dateList);
+        if (bookedList.size()==0) {
+            List<Client> clients = campsiteRepository.checkClientExisting(reservation.getClient());
+            if (clients.size()==0 ) {
+                long clientId = campsiteRepository.insertClient(reservation.getClient());
+                reservation.getClient().setId(clientId);
+                logger.debug("new client inserted, client id:" + clientId);
+            } else {
+                //TODO if client existing but name is different as database, do we need update client??
+                reservation.getClient().setId(clients.get(0).getId());
+            }
+
+            try {
+                int records = campsiteRepository.reserveDates(dateList);
+                if (logger.isDebugEnabled()) logger.debug("Total days:" + records + " reserved for client" + reservation.getClient().getName());
+                String reservationId =  campsiteRepository.createReservation(reservation);
+                if (logger.isDebugEnabled()) logger.debug("New Reservation created:" + reservationId);
+                reservation.setId(reservationId);
+            } catch (Exception e) {
+                logger.error("Error on the reservation:" + e);
+                throw new DataProcessException("Cannot add reservation; Some of dates have been reserved. Please try again");
+            }
         } else {
-            //TODO if client existing by name is different as database, do we need update client??
-            reservation.getClient().setId(clients.get(0).getId());
+            logger.error("Error on the reservation, campsite is no available for the input period.");
+            throw new DataProcessException("Error on the reservation, campsite is no available for the input period. Follow dates have been booked" + reservedDates(bookedList));
         }
 
-        try {
-            int records = campsiteRepository.reserveDates(dateList);
-            if (logger.isDebugEnabled()) logger.debug("Total days:" + records + " reserved for client" + reservation.getClient().getName());
-            String reservationId =  campsiteRepository.createReservation(reservation);
-            if (logger.isDebugEnabled()) logger.debug("New Reservation created:" + reservationId);
-            reservation.setId(reservationId);
-        } catch (Exception e) {
-            logger.error("Error on the reservation:" + e);
-            throw new DataProcessException("Cannot add reservation; Some of dates have been reserved. Please try again");
-        }
         return reservation;
     }
 
@@ -87,21 +90,31 @@ public class CampsiteServiceImpl implements CampsiteService{
             logger.error("Error on the reservation:" + e);
             throw new DataNotFoundException("Cannot found reservation, input reservation id is not existing.");
         }
-        if (oldReservation!=null) {
+        if (oldReservation!=null && reservationMarch(reservation, oldReservation)) {
             List<LocalDate> dateList = new ArrayList<>();
             long days = Duration.between(oldReservation.getArrival().atStartOfDay(), oldReservation.getDeparture().atStartOfDay() ).toDays();
             for (int i=0; i<=days; i++) {
                 dateList.add(reservation.getArrival().plusDays(i));
             }
             campsiteRepository.deleteDates(dateList);
-            reservation.setId(reservationId);
-            campsiteRepository.updateReservation(reservation);
+
             dateList.clear();
             days = Duration.between(reservation.getArrival().atStartOfDay(), reservation.getDeparture().atStartOfDay() ).toDays();
             for (int i=0; i<=days; i++) {
                 dateList.add(reservation.getArrival().plusDays(i));
             }
-            campsiteRepository.reserveDates(dateList);
+            List<LocalDate> bookedList =  campsiteRepository.verifyDates(dateList);
+            if (bookedList.size()==0) {
+                reservation.setId(reservationId);
+                campsiteRepository.updateReservation(reservation);
+                campsiteRepository.reserveDates(dateList);
+            } else {
+                logger.error("Error on the reservation, campsite is no available for the input period.");
+                throw new DataProcessException("Error on the reservation, campsite is no available for the input period. Follow dates have been booked" + reservedDates(bookedList));
+            }
+        } else {
+            throw new DataProcessException("Cannot update the reservation, please check your request again" );
+
         }
         return reservation;
     }
@@ -128,5 +141,15 @@ public class CampsiteServiceImpl implements CampsiteService{
             campsiteRepository.deleteReservation(reservationId);
         }
         return reservation;
+    }
+
+    private String reservedDates(List<LocalDate> dateList) {
+        return dateList.stream()
+                .map(n -> String.valueOf(n))
+                .collect(Collectors.joining(" ; ", "{", "}"));
+    }
+
+    private boolean reservationMarch(Reservation reservation, Reservation oldReservation) {
+        return oldReservation.getClient().getEmail().equalsIgnoreCase(reservation.getClient().getEmail());
     }
 }
